requirements.txt
absl-py==1.4.0
astunparse==1.6.3
blinker==1.6.2
boltons @ file:///C:/b/abs_707eo7c09t/croot/boltons_1677628723117/work
brotlipy==0.7.0
cachetools==5.3.1
certifi @ file:///C:/b/abs_36eb5mzhph/croot/certifi_1690232276943/work/certifi
cffi @ file:///C:/ci_310/cffi_1642682485096/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.6
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
conda @ file:///C:/b/abs_b5q92oum8h/croot/conda_1689272261421/work
conda-content-trust @ file:///tmp/build/80754af9/conda-content-trust_1617045594566/work
conda-package-handling @ file:///C:/ci/conda-package-handling_1649106011304/work
contourpy==1.1.0
cryptography @ file:///C:/ci/cryptography_1639414720302/work
cycler==0.11.0
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.41.1
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
h5py==3.9.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
keras==2.13.1
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.6
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
menuinst @ file:///C:/ci/menuinst_1631733438520/work
mkl-fft==1.3.6
mkl-random @ file:///C:/Users/dev-admin/mkl/mkl_random_1682977971003/work
mkl-service==2.4.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.74
opt-einsum==3.3.0
packaging @ file:///C:/b/abs_ed_kb9w6g4/croot/packaging_1678965418855/work
Pillow==9.4.0
pluggy @ file:///C:/ci/pluggy_1648024580010/work
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil==2.8.2
PyWavelets==1.4.1
pywin32==302
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.1
rsa==4.9
ruamel-yaml-conda @ file:///C:/ci/ruamel_yaml_1616016898638/work
ruamel.yaml @ file:///C:/b/abs_30ee5qbthd/croot/ruamel.yaml_1666304562000/work
ruamel.yaml.clib @ file:///C:/b/abs_aarblxbilo/croot/ruamel.yaml.clib_1666302270884/work
scikit-image==0.21.0
scipy==1.11.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tifffile==2023.7.18
toolz @ file:///C:/b/abs_cfvk6rc40d/croot/toolz_1667464080130/work
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm @ file:///opt/conda/conda-bld/tqdm_1647339053476/work
typing_extensions==4.5.0
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
Werkzeug==2.3.6
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
wrapt==1.15.0
zipp==3.16.2
